refactor: reducers, actions 디렉토리 구조 수정
- redux 폴더로 합침
